Destructure auth components
